+++ src/components/Common/Input/Select.tsx
@@ -1,17 +1,6 @@
 import React from "react"
 import classNames from "classnames"
 
-export interface SelectProps {
-    value?: string | number
-    onChange?: (e: React.ChangeEvent<HTMLSelectElement>) => void
-    rest?: any
-    label?: string
-    error?: string
-    options?: { value: string; label: string }[]
-    className?: string
-    disabled?: boolean
-}
-
 const Select = ({ onChange, value, rest, label, error, options = [], className, disabled }: SelectProps) => {
     return (
         <div className="kms_input-wrapper">
