+++ src/index.tsx
@@ -1,4 +1,5 @@
 import "./style.css"
 import Provider from "./context/ProviderContext"
 import { Master, SubMaster } from "./components"
-export { Master, SubMaster, Provider }
+import { ImageUpload, Input } from './components/Common'
+export { Master, SubMaster, Provider, ImageUpload, Input }
